refactor: Common library for automatically watching for changes and rebuilding The goal is to reuse this in both an upcoming `serve` command, and add the functionality to the `build` command.
